32 copies

+++ fuzzytools/boostraping.py
@@ -33,7 +33,7 @@ class BalancedCyclicBoostraping():
 			return
 		self.reset_counter()
 		self.class_names, counts = np.unique(self.l_classes, return_counts=True)
-		self.n = max(counts)*self.k_n
+		self.n = int(max(counts)*self.k_n)
 		self.l_objs_dict = {}
 		for c in self.class_names:
 			self.l_objs_dict[c] = [obj for obj,_c in zip(self.l_objs, self.l_classes) if _c==c]
